Optimized speed for huge archives like 250,000+ file .tar and .7z files

+++ src/Program.js
@@ -197,26 +197,25 @@ export class Program
 
 			// next we fix any filenames that contain UTF16 or other non-UTF8 characters, converting them to UTF8. This fixes problems with tree/readdir etc. because deno only supports UTF8 encodings
 			// can run `convmv --list` for a list of valid encoding names
-			xlog.trace`Program ${fg.orange(this.programid)} fixing filename encoding...`;
+			xlog.debug`Program ${fg.orange(this.programid)} fixing filename encoding...`;
 			const targetEncoding = (flags.filenameEncoding || (typeof this.filenameEncoding==="function" ? await this.filenameEncoding(r) : this.filenameEncoding)) || "windows-1252";
 			await runUtil.run("convmv", ["-r", "--qfrom", "--qto", "--notest", "-f", targetEncoding, "-t", "UTF-8", f.outDir.absolute]);
 
 			// delete various things
-			xlog.trace`Program ${fg.orange(this.programid)} deleting directory symlinks, empty files, 'special' files, and broken symlinks...`;
+			xlog.debug`Program ${fg.orange(this.programid)} deleting directory symlinks, empty files, 'special' files, and broken symlinks...`;
 			await runUtil.run("find", [f.outDir.absolute, "-type", "l", "-xtype", "d", "-delete"]);		// delete directory symlinks, which are dangerous as they could point outside of the output directory
 			await runUtil.run("find", [f.outDir.absolute, "-type", "b,c,p,s", "-delete"]);				// delete special files like block, named pipes, sockets, etc
 			await runUtil.run("find", [f.outDir.absolute, "-type", "f", "-empty", "-delete"]);			// delete empty files
 			await runUtil.run("find", [f.outDir.absolute, "-xtype", "l", "-delete"]);					// delete broken symlinks. make sure we do this last just in case a symlink becomes broken because we have deleted the file it points to above
 
 			// now find any new files on disk in the output dir that we don't yet
-			xlog.trace`Program ${fg.orange(this.programid)} locating new files...`;
+			xlog.debug`Program ${fg.orange(this.programid)} locating new files...`;
 			const newFilePaths = (await fileUtil.tree(f.outDir.absolute, {nodir : true})).subtractAll([...(f.files.output || []), ...(f.files.prev || []), ...(f.files.new || [])].map(v => v.absolute));
 
-			xlog.trace`Program ${fg.orange(this.programid)} doing first pass on ${newFilePaths.length} new files...`;
-			await newFilePaths.parallelMap(async newFilePath =>
+			xlog.debug`Program ${fg.orange(this.programid)} sanity filtering ${newFilePaths.length.toLocaleString()} new files...`;
+			const newDexFiles = (await newFilePaths.parallelMap(async newFilePath =>
 			{
 				const newFileRel = path.relative(f.outDir.absolute, newFilePath);
-
 				const newFile = await DexFile.create({root : f.root, absolute : newFilePath});
 
 				if(newFile.size>xu.GB)
@@ -260,22 +259,24 @@ export class Program
 
 				// check to see if the resulting file is identical to our input file (sample/executable/exe/Oidata)
 				// we always check if we have just 1 output file, otherwise we only check if checkForDups is set
-				if(!this.allowDupOut && (this.checkForDups || newFilePaths.length===1) && await fileUtil.areEqual(newFilePath, f.input.absolute))
+				if(f.input.size<xu.MB*25 && !this.allowDupOut && (this.checkForDups || newFilePaths.length===1) && await fileUtil.areEqual(newFilePath, f.input.absolute))
 				{
 					xlog.warn`Program ${fg.orange(this.programid)} deleting output file ${newFileRel} due to being identical to input file ${f.input.pretty()}`;
 					await fileUtil.unlink(newFilePath);
 					return;
 				}
 
-				// add our new file to our fileset
-				await f.add("new", newFile);
-			});
+				return newFile;
+			})).filter(v => !!v);
+
+			xlog.debug`Adding ${newDexFiles.length.toLocaleString()} new files to file set...`;
+			await f.addAll("new", newDexFiles);
 		}
 
 		// sort the filenames by depth and then by filename
 		if(f.files.new)
 		{
-			xlog.debug`Program ${fg.orange(this.programid)} located ${f.files.new.length} new files...`;
+			xlog.debug`Program ${fg.orange(this.programid)} located ${f.files.new.length.toLocaleString()} new files...`;
 			f.files.new.sortMulti([file => file.rel.split("/").length, file => file.base]);
 		}
